// src/redux/actions/action-types.ts
export enum ActionType {
  RESET_STATE = "initial state",
  ADD_TODO = "add todo",
  DELETE_TODO = "delete todo",
  COMPLETE_TODO = "complete todo",
  UPDATE_TODO_TEXT = "update todo text",
  UPDATE_EDIT_STATUS = "update edit status",
  UPDATE_FILTER = "update filter",
  FILTER_TODOS = "filter todos",
}